# public/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BordApp API",
    "description": "Документація до API бордової дошки",
    "contact": {
      "email": "support@bord-app.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://board.filkx.com",
      "description": "Основний сервер"
    }
  ],
  "paths": {
...
    "/api/v1/user/show": {
      "get": {
        "summary": "Отримати профіль користувача",
        "operationId": "d1b09ef443e0e5090861830428b52988",
        "responses": {
          "200": {
            "description": "Профіль користувача",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          },
          {
            "OAuth2": []
          }
        ]
      }
    },
    "/api/v1/user": {
      "put": {
        "summary": "Оновити профіль користувача",
        "operationId": "a42e38e4fe243e3394defedfd548c6d8",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Профіль оновлено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          },
          {
            "OAuth2": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ProfileUpdateRequest": {
        "required": ["first_name", "name", "last_name", "email"],
        "properties": {
          "first_name": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "type": "object"
      },
...
      "Profile": {
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "properties": {
              "first": {
                "type": "string"
              },
              "last": {
                "type": "string"
              }
            },
            "type": "object"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "properties": {
              "number": {
                "type": "string"
              },
              "varified": {
                "type": "boolean"
              }
            },
            "type": "object"
          },
          "locale": {
            "type": "string"
          },
          "avatar_url": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "bearerFormat": "JWT",
        "scheme": "bearer"
      }
    }
  },
...
}